backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
celery
redis
requests
beautifulsoup4
lxml
aiohttp
pymupdf
